move update profile photo to edit profile section
