+++ src/components/users/PublicProfile.tsx
@@ -1,13 +1,11 @@
 import { MouseEventHandler } from 'react'
 import Link from 'next/link'
-import { PencilIcon } from '@heroicons/react/24/solid'
 
 import { UserPublicProfile } from '../../js/types/User'
 import EditProfileButton from './EditProfileButton'
 import ImportFromMtnProj from './ImportFromMtnProj'
 import APIKeyCopy from './APIKeyCopy'
 import usePermissions from '../../js/hooks/auth/usePermissions'
-import { BaseProfilePhotoUploader } from '../media/BaseUploader'
 import { DefaultLoader } from '../../js/sirv/util'
 
 interface PublicProfileProps {
@@ -25,7 +23,11 @@ export default function PublicProfile ({ userProfile }: PublicProfileProps): JSX
   }
   return (
     <section className='mx-auto max-w-screen-sm px-4 md:px-0 md:grid md:grid-cols-3'>
-      <ProfileImage avatar={avatar ?? ''} />
+      <img
+        className='object-cover w-24 h-24 rounded-full'
+        src={avatar != null && avatar !== '' ? DefaultLoader({ src: avatar, width: 200 }) : ''}
+        alt='Profile Photo'
+      />
       <div className='md:col-span-2 text-medium text-primary '>
         <div className='flex flex-row items-center gap-x-2 max-w-xs'>
           <div className='md:text-2xl font-bold mr-4 truncate'>
@@ -102,23 +104,3 @@ export const ProfileATag = ({ uid, className = ProfileATagStyle }: ProfileATagPr
   </Link>)
 
 const ProfileATagStyle = 'text-primary font-bold hover:underline'
-
-export const ProfileImage = ({ avatar }: { avatar: string }): JSX.Element => {
-  return (
-    <div
-      className='hidden md:block pr-5 relative'
-      aria-label='Profile photo section'
-    >
-      <div className='relative inline-block'>
-        <BaseProfilePhotoUploader className='absolute bottom-1 right-1  bg-gray-800 bg-opacity-75 p-2 rounded-full transition-opacity duration-200 hover:bg-opacity-100'>
-          <PencilIcon className='w-4 h-4 text-white' />
-        </BaseProfilePhotoUploader>
-        <img
-          className='object-cover w-24 h-24 rounded-full'
-          src={DefaultLoader({ src: avatar, width: 200 })}
-          alt='Profile Photo'
-        />
-      </div>
-    </div>
-  )
-}
